requirements.txt
# Core app
streamlit==1.48.1
streamlit-js-eval==0.1.7
streamlit-mic-recorder==0.0.8

# LLM bindings (CPU build by default)
llama-cpp-python

# TTS + audio
kokoro==0.9.4
soundfile==0.13.1
torchaudio
torch
